Normalize paths within fs.FS
The shrt command provides an fs.FS rooted at "/" to the ShrtHandler.
Files are accessed within the FS as relative paths. As a consequence,
paths passed into the ShrtHandler Config must be relative to the root
of the file system. This is equivalent to saying paths must be
absolute, but must trim the leading "/".

This commit modifies the env subcommand to satisfy this requirement.
Paths provided to override SHRT_DBPATH must now be absolute.